docs: add monospace font
Let's add another CSS variable to our collection of theming tokens in order to be able to use the same monospace font family across plugins. I'd suggest going with Roboto Mono in the community distribution.

If we adopt this change, we should then also update the ACE editor component inside our wizards to make use of that font.